+++ trouble_code_reader.h
@@ -15,7 +15,7 @@ extern TROUBLE_CODE master_trouble_list[];
 int display_trouble_codes(void);
 int handle_read_codes(char *, int);
 void ready_trouble_codes(void);
-void printTroubleCodes(char *buf, size_t bufSize);
+void printTroubleCodes(char *buf, unsigned long bufSize);
 void initializeUnknownList();
 void destroyUnknownList();
